move in squareset
