+++ src/chess/mod.rs
@@ -1,4 +1,5 @@
 pub mod board;
 pub mod chessmove;
 pub mod piece;
+pub mod squareset;
 pub mod types;
