+++ src/chess/squareset.rs
