+++ src/datagen/dataformat/marlinformat.rs
@@ -1,10 +1,8 @@
-use crate::{
-    chess::{
-        board::{Board, GameOutcome},
-        piece::{Colour, Piece, PieceType},
-        types::{Rank, Square},
-    },
+use crate::chess::{
+    board::{Board, GameOutcome},
+    piece::{Colour, Piece, PieceType},
     squareset::SquareSet,
+    types::{Rank, Square},
 };
 
 const UNMOVED_ROOK: u8 = 6; // one higher than the max piecetype enum value
