+++ src/lookups.rs
@@ -1,6 +1,9 @@
 #![allow(clippy::cast_possible_truncation)]
 
-use crate::{chess::types::Square, rng::XorShiftState, squareset::SquareSet};
+use crate::{
+    chess::{squareset::SquareSet, types::Square},
+    rng::XorShiftState,
+};
 
 /// Implements a C-style for loop, for use in const fn.
 #[macro_export]
@@ -112,9 +115,10 @@ mod tests {
 
     #[test]
     fn python_chess_validation() {
-        use crate::chess::types::Square;
-        use crate::lookups::{get_king_attacks, get_knight_attacks};
-        use crate::squareset::SquareSet;
+        use crate::{
+            chess::{squareset::SquareSet, types::Square},
+            lookups::{get_king_attacks, get_knight_attacks},
+        };
         // testing that the attack squaresets match the ones in the python-chess library,
         // which are known to be correct.
         assert_eq!(
